--- a/.gitignore
+++ b/.gitignore
@@ -14,3 +14,4 @@ tmp/
 coverage
 storybook-static/
 packages/icons/dist/
+.cache-loader/