--- a/package.json
+++ b/package.json
@@ -82,6 +82,7 @@
     "@types/storybook__addon-actions": "3.4.1",
     "@types/storybook__addon-info": "3.4.2",
     "@types/storybook__addon-knobs": "3.4.1",
+    "cache-loader": "2.0.1",
     "commander": "2.19.0",
     "cz-conventional-changelog": "2.1.0",
     "enzyme": "3.7.0",