--- a/webpack/webpack.base.js
+++ b/webpack/webpack.base.js
@@ -1,5 +1,6 @@
 const path = require("path");
 const webpack = require("webpack");
+const ForkTsCheckerWebpackPlugin = require("fork-ts-checker-webpack-plugin");
 
 const webpackBase = {
   getDevTool: () => {
@@ -11,7 +12,11 @@ const webpackBase = {
       {
         test: /\.(js|ts|tsx?)$/,
         exclude: /node_modules/,
-        use: ["cache-loader", "ts-loader", "react-docgen-typescript-loader"]
+        use: [
+          "cache-loader",
+          { loader: "ts-loader", options: { transpileOnly: true } },
+          "react-docgen-typescript-loader"
+        ]
       },
       {
         test: /\.(eot|ttf|woff|woff2|svg|png|gif|jpe?g)$/,
@@ -25,7 +30,8 @@ const webpackBase = {
     const plugins = [
       new webpack.DefinePlugin({
         "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV)
-      })
+      }),
+      new ForkTsCheckerWebpackPlugin()
     ];
 
     if (process.env.NODE_ENV === "production") {