@AmrAbdelrazik 2019-02-21
hey @pierrebeitz is that compile/engineering workflow performance improvement only or do customers (DC/OS UI users) see benefits as well?