@pierrebeitz 2019-02-21
@jaxesn i have an unfinished spike for dcos-ui (it's a little complicated as it reintroduces type checking as well). 
will look into the protoss-setup afterwards.

@AmrAbdelrazik these changes are for engineers only.